# api/openapi.yaml
openapi: "3.1.0"
info:
  title: OurFamRoots API
  version: "1.0.0"
  description: |
    ## OurFamRoots REST API — v1

    A multi-tenant genealogy platform API following REST conventions.

    ### Versioning
    URL-path versioning: all v1 endpoints are prefixed `/api/v1`.
    Breaking changes increment the major version (`/api/v2`).
    Minor/patch changes are backward-compatible and do not change the URL.

    ### Authentication
    Two tokens are issued on login:
    - **Access token** — short-lived JWT (15 min), passed as `Authorization: Bearer <token>`
    - **Refresh token** — long-lived (30 days), stored as an `HttpOnly SameSite=Strict` cookie

    ### Multi-tenancy
    Every request is scoped to a tenant derived from the JWT `tenant_id` claim.
    Tenant is never supplied by the caller — it is always resolved server-side from the token.

    ### Rate limiting
    All endpoints return rate-limit headers:
    - `X-RateLimit-Limit` — requests allowed in the current window
    - `X-RateLimit-Remaining` — requests remaining in the current window
    - `X-RateLimit-Reset` — Unix timestamp when the window resets
    - `Retry-After` — seconds to wait (only present on 429 responses)

    Default limits by plan:
    | Plan        | Requests/min | Burst |
    |-------------|-------------|-------|
    | FREE        | 60          | 10    |
    | BASIC       | 300         | 30    |
    | PREMIUM     | 1000        | 100   |
    | PROFESSIONAL| 5000        | 500   |

...
    ```
...
    ```
...
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT access token issued on login. Expires in 15 minutes.
        Payload includes: sub (user_id), tenant_id, email, roles, plan.

    CookieAuth:
      type: apiKey
      in: cookie
      name: refresh_token
      description: >
        HttpOnly SameSite=Strict cookie holding the refresh token.
        Used only on POST /auth/refresh and POST /auth/logout.

  # ──────────────────────────────────────────────────────────
  # COMMON PARAMETERS
  # ──────────────────────────────────────────────────────────
...
  schemas:
...
    Error:
      type: object
      required: [type, title, status, requestId]
      properties:
        type:
          type: string
          format: uri
          example: "https://ourfamroots.com/errors/validation-error"
        title:
          type: string
          example: "Validation Error"
        status:
          type: integer
          example: 422
        detail:
          type: string
          example: "The request body contains invalid fields."
        instance:
          type: string
          example: "/api/v1/trees/abc/persons"
        errors:
          type: array
          items:
            $ref: "#/components/schemas/FieldError"
        requestId:
          type: string
          example: "req_01hzab3k7x"
    FieldError:
      type: object
      required: [field, code, message]
      properties:
        field:
          type: string
          example: "sex"
        code:
          type: string
          example: "INVALID_ENUM"
        message:
          type: string
          example: "Must be one of: MALE, FEMALE, OTHER, UNKNOWN"

    # ── Pagination ──────────────────────────────────────────
...
    TokenPair:
      type: object
      required: [accessToken, tokenType, expiresIn, user]
      properties:
        accessToken:
          type: string
          description: Short-lived JWT (15 min)
        tokenType:
          type: string
          enum: [Bearer]
        expiresIn:
          type: integer
          example: 900
          description: Access token TTL in seconds
        user:
          $ref: "#/components/schemas/UserProfile"

    # ── User ────────────────────────────────────────────────
    UserProfile:
      type: object
      required: [id, email, displayName, isVerified, plan, createdAt]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        displayName:
          type: string
          maxLength: 100
        avatarUrl:
          type: string
          format: uri
          nullable: true
        locale:
          type: string
          example: "en"
        timezone:
          type: string
          example: "America/New_York"
        isVerified:
          type: boolean
        plan:
          type: string
          enum: [FREE, BASIC, PREMIUM, FAMILY, PROFESSIONAL]
        createdAt:
          type: string
          format: date-time
...
  responses:
    400BadRequest:
      description: Bad Request — malformed JSON or missing required fields
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    429TooManyRequests:
      description: Too Many Requests — rate limit exceeded
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds to wait before retrying
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"

# ──────────────────────────────────────────────────────────────
# GLOBAL SECURITY (override per-operation where needed)
# ──────────────────────────────────────────────────────────────
security:
  - BearerAuth: []

# ──────────────────────────────────────────────────────────────
# TAGS
# ──────────────────────────────────────────────────────────────
...
paths:
...
  /auth/login:
    post:
      tags: [Authentication]
      operationId: login
      summary: Authenticate with email and password
      description: |
        Returns an access token (JSON body) and sets an `HttpOnly` refresh token
        cookie. Login is blocked with `403 account-not-verified` until the
        user's email has been verified (either via the link in the registration
        email or manually by an Admin).
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                rememberMe:
                  type: boolean
                  default: false
                  description: Extends refresh token TTL from 30 days to 90 days
      responses:
        "200":
          description: Login successful
          headers:
            Set-Cookie:
              schema:
                type: string
              description: "HttpOnly SameSite=Lax refresh_token cookie"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPair"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Email not verified (`type` = `account-not-verified`)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "423":
          description: Account locked after too many failed attempts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          $ref: "#/components/responses/429TooManyRequests"
...
  /auth/oauth/{provider}:
    get:
      tags: [Authentication]
      operationId: oauthRedirect
      summary: Redirect to OAuth provider authorisation page
      security: []
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
            enum: [google, facebook, apple]
        - name: redirectUri
          in: query
          required: true
          schema:
            type: string
            format: uri
      responses:
        "302":
          description: Redirect to provider
        "400":
          $ref: "#/components/responses/400BadRequest"
  /auth/oauth/{provider}/callback:
    get:
      tags: [Authentication]
      operationId: oauthCallback
      summary: Handle OAuth provider callback
      security: []
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
            enum: [google, facebook, apple]
        - name: code
          in: query
          required: true
          schema:
            type: string
        - name: state
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OAuth login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPair"
        "400":
          $ref: "#/components/responses/400BadRequest"
        "409":
          description: Email already registered with a different provider

  # ============================================================
  # USERS
  # ============================================================
...